Fix confirmation logic in handleDelete function in App.jsx to prevent deletion if user cancels

+++ client/src/App.jsx
@@ -34,10 +34,12 @@ function App() {
     const isConfirmed = window.confirm(
       "Are you sure you want to delete this user?"
     );
-    await axios.delete(`http://localhost:3030/users/${id}`).then((res) => {
-      setUsers(res.data);
-      setFilterusers(res.data);
-    });
+    if (!isConfirmed) {
+      await axios.delete(`http://localhost:3030/users/${id}`).then((res) => {
+        setUsers(res.data);
+        setFilterusers(res.data);
+      });
+    }
   };
 
   return (
